Nodes are only interactable when enabled, fixing large mesh performance issues; UI updates
