+++ src/components/al-console/readme.md
@@ -24,7 +24,7 @@
 
 ### Used by
 
- - [al-control-panel](..\al-control-panel)
+ - [al-control-panel](../al-control-panel)
 
 ### Depends on
