+++ src/components/al-graph-settings/ContentStrings.ts
@@ -1,0 +1,31 @@
+import { I18nJson } from "../../interfaces/I18nJson";
+
+export interface ContentStrings extends I18nJson {
+  axial?: string;
+  boundingBoxEnabled?: string;
+  brightness?: string;
+  clay?: string;
+  contrast?: string;
+  controlsType?: string;
+  coronal?: string;
+  default?: string;
+  displayMode?: string;
+  graphEnabled?: string;
+  material?: string;
+  meters?: string;
+  millimeters?: string;
+  normals?: string;
+  orbit?: string;
+  orientation?: string;
+  recenter?: string;
+  saggital?: string;
+  sliceIndex?: string;
+  slices?: string;
+  trackball?: string;
+  units?: string;
+  volume?: string;
+  volumeSteps?: string;
+  volumeStepsHighEnabled?: string;
+  wireframe?: string;
+  xray?: string;
+}
