+++ src/components/al-graph-settings/al-graph-settings.i18n.en.json
@@ -1,0 +1,6 @@
+{
+  "graphEnabled": "Enable Nodes",
+  "units": "Measurement Units",
+  "meters": "Meters",
+  "millimeters": "Millimeters"
+}
