+++ src/components/al-node-list/al-node-list.i18n.en.json
@@ -1,3 +1,3 @@
 {
-  "graphEmpty": "The graph is currently empty"
+  "graphEmpty": "No nodes have been placed."
 }
