+++ src/components/al-node-list/readme.md
@@ -24,7 +24,7 @@
 
 ### Used by
 
- - [al-graph-editor](..\al-graph-editor)
+ - [al-graph-editor](../al-graph-editor)
 
 ### Depends on
