+++ src/components/al-settings/al-settings.tsx
@@ -1,11 +1,6 @@
 import { Component, Event, EventEmitter, h, Prop } from "@stencil/core";
 import AlertIcon from "../../assets/svg/alert.svg";
-import BoundingBoxIcon from "../../assets/svg/bounding-box-2.svg";
-import ObjectIcon from "../../assets/svg/object-alone.svg";
-import OrbitCameraIcon from "../../assets/svg/orbit_camera.svg";
-import RecenterIcon from "../../assets/svg/recenter.svg";
-import RotateObjectIcon from "../../assets/svg/rotate_object.svg";
-import { ControlsType, Orientation, Units } from "../../enums";
+import { Orientation } from "../../enums";
 import { DisplayMode } from "../../enums/DisplayMode";
 import i18n from "./al-settings.i18n.en.json";
 import { ContentStrings } from "./ContentStrings";
@@ -18,63 +13,31 @@ import { ContentStrings } from "./ContentStrings";
 export class AlSettings {
   private _contentStrings: ContentStrings = i18n;
 
-  @Event() public boundingBoxEnabledChanged: EventEmitter;
-  @Event() public controlsTypeChanged: EventEmitter;
   @Event() public displayModeChanged: EventEmitter;
-  // @Event() public materialChanged: EventEmitter;
   @Event() public orientationChanged: EventEmitter;
-  @Event() public recenter: EventEmitter;
   @Event() public slicesIndexChanged: EventEmitter;
   @Event() public slicesBrightnessChanged: EventEmitter;
   @Event() public slicesContrastChanged: EventEmitter;
-  @Event() public unitsChanged: EventEmitter;
-  @Event() public graphEnabledChanged: EventEmitter;
   @Event() public volumeBrightnessChanged: EventEmitter;
   @Event() public volumeContrastChanged: EventEmitter;
   @Event() public volumeStepsChanged: EventEmitter;
   @Event() public volumeStepsHighEnabledChanged: EventEmitter;
 
-  @Prop({ mutable: true }) public boundingBoxEnabled: boolean;
-  @Prop({ mutable: true }) public controlsType: ControlsType;
   @Prop({ mutable: true }) public displayMode: DisplayMode;
-  @Prop({ mutable: true }) public graphEnabled: boolean;
-  @Prop({ mutable: true }) public graphVisible: boolean;
-  // @Prop({ mutable: true }) public material: Material = Material.DEFAULT;
   @Prop({ mutable: true }) public orientation: Orientation;
   @Prop({ mutable: true }) public slicesIndex: number;
   @Prop({ mutable: true }) public slicesMaxIndex: number;
   @Prop({ mutable: true }) public slicesBrightness: number; // window center
   @Prop({ mutable: true }) public slicesContrast: number; // window width
-  @Prop({ mutable: true }) public units: Units;
   @Prop({ mutable: true }) public volumeBrightness: number; // window center
   @Prop({ mutable: true }) public volumeContrast: number; // window width
   @Prop({ mutable: true }) public volumeSteps: number;
   @Prop({ mutable: true }) public volumeStepsHighEnabled: boolean;
 
-  private _boundingBoxEnabled(enabled: boolean) {
-    this.boundingBoxEnabled = enabled;
-    this.boundingBoxEnabledChanged.emit(enabled);
-  }
-
-  private _controlsType(controlsType: ControlsType) {
-    this.controlsType = controlsType;
-    this.controlsTypeChanged.emit(controlsType);
-  }
-
   private _displayMode(displayMode: DisplayMode) {
     this.displayMode = displayMode;
     this.displayModeChanged.emit(displayMode);
   }
-
-  private _graphEnabled(enabled: boolean) {
-    this.graphEnabled = enabled;
-    this.graphEnabledChanged.emit(enabled);
-  }
-
-  // private _material(material: Material) {
-  //   this.material = material;
-  //   this.materialChanged.emit(material);
-  // }
 
   private _orientation(orientation: Orientation) {
     this.orientation = orientation;
@@ -94,27 +57,6 @@ export class AlSettings {
   private _slicesContrast(contrast: number) {
     this.slicesContrast = contrast;
     this.slicesContrastChanged.emit(contrast);
-  }
-
-  private _switchBoundingBoxEnabled() {
-    if (this.boundingBoxEnabled) {
-      this._boundingBoxEnabled(false);
-    } else {
-      this._boundingBoxEnabled(true);
-    }
-  }
-
-  private _switchControls() {
-    if (this.controlsType === ControlsType.ORBIT) {
-      this._controlsType(ControlsType.TRACKBALL);
-    } else if (this.controlsType === ControlsType.TRACKBALL) {
-      this._controlsType(ControlsType.ORBIT);
-    }
-  }
-
-  private _units(units: Units) {
-    this.units = units;
-    this.unitsChanged.emit(this.units);
   }
 
   private _volumeBrightness(brightness: number) {
@@ -139,131 +81,6 @@ export class AlSettings {
     } else {
       this._volumeSteps(this.volumeSteps - 0.2);
     }
-  }
-
-  public renderControlsTypeSelect() {
-    let cameraIcon: string;
-    let cameraLabel: string;
-    if (this.controlsType === ControlsType.ORBIT) {
-      cameraIcon = OrbitCameraIcon;
-      cameraLabel = this._contentStrings.orbit;
-    } else if (this.controlsType === ControlsType.TRACKBALL) {
-      cameraIcon = RotateObjectIcon;
-      cameraLabel = this._contentStrings.rotate;
-    }
-    let boundingBoxEnabledIcon: string;
-    if (this.boundingBoxEnabled) {
-      boundingBoxEnabledIcon = BoundingBoxIcon;
-    } else {
-      boundingBoxEnabledIcon = ObjectIcon;
-    }
-    return (
-      <div
-        style={{
-          "margin-top": "10px",
-          "text-align": "center"
-        }}
-      >
-        <ion-button
-          style={{
-            width: "28%",
-            height: "45px",
-            "margin-left": "5px",
-            "margin-right": "5px"
-          }}
-          size="small"
-          onClick={() => {
-            this._switchControls();
-          }}
-        >
-          <div
-            style={{
-              "font-size": "10px",
-              color: "white",
-              "margin-bottom": "2px"
-            }}
-          >
-            <ion-icon
-              style={{
-                "min-width": "20px",
-                "min-height": "20px",
-                "margin-bottom": "2px"
-              }}
-              src={cameraIcon}
-              title={cameraLabel}
-            />
-            <br />
-            {cameraLabel}
-          </div>
-        </ion-button>
-
-        <ion-button
-          style={{
-            width: "28%",
-            height: "45px",
-            "margin-left": "5px",
-            "margin-right": "5px"
-          }}
-          size="small"
-          onClick={() => {
-            this.recenter.emit();
-          }}
-        >
-          <div
-            style={{
-              "font-size": "10px",
-              color: "white",
-              "margin-bottom": "2px"
-            }}
-          >
-            <ion-icon
-              style={{
-                "min-width": "20px",
-                "min-height": "20px",
-                "margin-bottom": "2px"
-              }}
-              src={RecenterIcon}
-              title={this._contentStrings.recenter}
-            />
-            <br />
-            {this._contentStrings.recenter}
-          </div>
-        </ion-button>
-
-        <ion-button
-          style={{
-            width: "28%",
-            height: "45px",
-            "margin-left": "5px",
-            "margin-right": "5px"
-          }}
-          size="small"
-          onClick={() => {
-            this._switchBoundingBoxEnabled();
-          }}
-        >
-          <div
-            style={{
-              "font-size": "10px",
-              color: "white",
-              "margin-bottom": "2px"
-            }}
-          >
-            <ion-icon
-              style={{
-                "min-width": "20px",
-                "min-height": "20px",
-                "margin-bottom": "2px"
-              }}
-              src={boundingBoxEnabledIcon}
-              title={this._contentStrings.bounds}
-            />
-            <br />
-            {this._contentStrings.bounds}
-          </div>
-        </ion-button>
-      </div>
-    );
   }
 
   public renderDisplayModeToggle() {
@@ -301,68 +118,6 @@ export class AlSettings {
     }
 
     return null;
-  }
-
-  public renderGraphEnabled() {
-    if (this.graphVisible) {
-      return (
-        <div>
-          <ion-list-header
-            style={{
-              color: "var(--al-item-color)",
-              "border-width": "1px 0 0 0",
-              "border-color": "var(--ion-list-header-border-color)",
-              "border-style": "solid",
-              "margin-top": "10px"
-            }}
-          >
-            <span>Measure and Annotate</span>
-          </ion-list-header>
-          <ion-item
-            style={{
-              display: "var(--graph-enabled-display, block)"
-            }}
-          >
-            <span title={this._contentStrings.graphEnabled}>
-              {this._contentStrings.graphEnabled}
-            </span>
-            <ion-toggle
-              slot="end"
-              checked={this.graphEnabled}
-              onIonChange={e => this._graphEnabled(e.detail.checked)}
-            />
-          </ion-item>
-          <ion-item
-            style={{
-              display: "var(--units-display, block)"
-            }}
-          >
-            <span title={this._contentStrings.units}>Units</span>
-            <select
-              slot="end"
-              onChange={e =>
-                this._units((e.srcElement as HTMLSelectElement).value as Units)
-              }
-            >
-              <option
-                selected={this.units === Units.MILLIMETERS}
-                value={Units.MILLIMETERS}
-              >
-                {this._contentStrings.millimeters}
-              </option>
-              <option
-                selected={this.units === Units.METERS}
-                value={Units.METERS}
-              >
-                {this._contentStrings.meters}
-              </option>
-            </select>
-          </ion-item>
-        </div>
-      );
-    } else {
-      return null;
-    }
   }
 
   public renderHiResEnabled() {
@@ -580,10 +335,8 @@ export class AlSettings {
           "overflow-x": "hidden"
         }}
       >
-        {this.renderControlsTypeSelect()}
         {this.renderDisplayModeToggle()}
         {this.renderOptions()}
-        {this.renderGraphEnabled()}
       </div>
     );
   }
