+++ src/components/al-tabs/readme.md
@@ -59,7 +59,7 @@ Type: `Promise<boolean>`
 
 ### Used by
 
- - [al-control-panel](..\al-control-panel)
+ - [al-control-panel](../al-control-panel)
 
 ### Graph
 ```mermaid
