+++ src/components/al-view-controls/ContentStrings.ts
@@ -1,0 +1,11 @@
+import { I18nJson } from "../../interfaces/I18nJson";
+
+export interface ContentStrings extends I18nJson {
+  boundingBoxEnabled?: string;
+  controlsType?: string;
+  default?: string;
+  graphEnabled?: string;
+  orbit?: string;
+  recenter?: string;
+  trackball?: string;
+}
